pyproject.toml
[tool.isort]
profile = "black"

[tool.coverage.run]
plugins = [
    'django_coverage_plugin',
]

[tool.poetry]
name = "simple-django"
version = "0.1.0"
description = "Simple Django project boilerplate"
authors = ["Josh Michael Karamuth <michael@confuzeus.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "simple_django"}]

[tool.poetry.dependencies]
python = "^3.10"
argon2-cffi = "^21.3.0"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
pymemcache = "^4.0.0"
django = "^4.1.6"
django-extensions = "^3.2.1"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.2"}
celery = "^5.3.1"
django-cors-headers = "^4.2.0"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.5"
factory-boy = "^3.2.1"
coverage = "^7.1.0"
django-coverage-plugin = "^3.0.0"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"


[tool.poetry.group.prod.dependencies]
gunicorn = "^20.1.0"
psycopg2 = "^2.9.5"
django-anymail = {extras = ["amazon-ses"], version = "^9.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
